requirements.txt
aiogram
flask
google-generativeai
aiohttp
gunicorn